backend/requirements.txt
djangorestframework
psycopg2
apsheduler
django-apscheduler
django-cors-headers
django
gunicorn